+++ tests/Forc.toml
@@ -10,13 +10,13 @@ members = [
   "./src/signed_integers/signed_i128",
   "./src/signed_integers/signed_i256",
   "./src/string",
-  "./src/unsigned_integers/ufp64_div_test",
-  "./src/unsigned_integers/ufp64_exp_test",
-  "./src/unsigned_integers/ufp64_mul_test",
-  "./src/unsigned_integers/ufp64_mul_test",
-  "./src/unsigned_integers/ufp64_pow_test",
-  "./src/unsigned_integers/ufp64_root_test",
-  "./src/unsigned_integers/ufp64_test",
-  "./src/unsigned_integers/ufp128_div_test",
-  "./src/unsigned_integers/ufp128_test",
+  "./src/unsigned_numbers/ufp64_div_test",
+  "./src/unsigned_numbers/ufp64_exp_test",
+  "./src/unsigned_numbers/ufp64_mul_test",
+  "./src/unsigned_numbers/ufp64_mul_test",
+  "./src/unsigned_numbers/ufp64_pow_test",
+  "./src/unsigned_numbers/ufp64_root_test",
+  "./src/unsigned_numbers/ufp64_test",
+  "./src/unsigned_numbers/ufp128_div_test",
+  "./src/unsigned_numbers/ufp128_test",
 ]
