+++ tests/src/unsigned_numbers/ufp64_mul_test/src/main.sw
